Refactor and simplify nspace route shims
